fix(security): resolve all CodeQL code scanning alerts
